[CodeGen]Encode liveness for copy used MO after virtRegRewriter.
As for the Greedy RA, the virtRegRewriter pass is the last place
that holds livenes info, even at subregister level. So, now that
information can be extracted and encoded on COPY instruction.

This information for COPY  can later be used (as long as liveness info remains valid) to identify partially live regsiters precisely.